Fix getting from a nested structure
Looks like there was a problem where the Struct case for handling struct fields. It properly
attempts a conversion to a Value wrapping the correct type of the struct field, however it does
not unwrap the Value. This causes nested structs to get returned as a Value wrapping a struct.
Subsequent attempts to access fields of the nested structure, would fail because the value being
handled is the wrapper instead of the bare value.

To fix this, we just ask for the Interface() of the Value before returning, thereby unwrapping the
nested struct.

Also added a test that has a struct in a struct in a struct. It works!