Handle arrays with non-interface{} element type
The type assertion that was originally in place caused arrays with element types that were not
empty interface to fail the type assertion. That left those types of arrays unhandled, resulting in
nil return. This fix handles that case by using the reflect.Value wrapper to do the type checking
and indexing. Also add a test that tests out an array of int pointers.