Handle nil values on indexing operations
Hitting a nil value previously would cause a crash in the experiment run. This happens in both the
case of the value being in a map, and in the case where the value is part of a struct. This fix causes
the code to instead return a nil.
Unfortunately this is not quite ideal in the case of maps to numerical values.  In a normal go map,
maps with numerical values apparently return a 0 when the requested key is not present. There
doesn't seem to be any simple way to maintain parity with that behavior (though it does seem kind
of odd, looking at it from Java colored glasses), so a concession is made to just return a nil in that
case.